% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brass_logits.R
\name{brass_logit_model}
\alias{brass_logit_model}
\title{Brass Logit Model}
\usage{
brass_logit_model(data, qx_col, age_col, standard, standards_data = NULL, ...)
}
\arguments{
\item{data}{A data frame containing the observed probabilities (qx).}

\item{qx_col}{A string specifying the column name for the observed probabilities (qx).}

\item{age_col}{A string specifying the column name for age groups.}

\item{standard}{A string specifying the standard to use ("African" or "GeneralUN").}

\item{standards_data}{A data frame containing the standard logits (must include columns for "Age", standard probabilities, and logits).}

\item{...}{Additional arguments passed to \code{lm()} for fitting the linear model.}
}
\value{
A data frame containing the original data and the predicted qx values, with predictions for missing qx.
}
\description{
This function fits a Brass Logit Model based on user-specified standards.
The function uses the linear transformation property of the Brass Logit Model.
}
\examples{
# Example data with NA for some qx values
observed_data <- data.frame(
  age = c(0, 1, 5, 10, 15, 20, 25, 30, 35, 40, 45, 50),
  qx = c(0.1, 0.05, 0.02, NA, 0.005, NA, 0.002, 0.0015, NA, 0.0008, 0.0006, 0.0005)
)

brass_logit_model(
  data = observed_data,
  qx_col = "qx",
  age_col = "age",
  standard = "African"
)


}
