% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/child_mortality.R
\name{dem.chm}
\alias{dem.chm}
\title{Calculate Child, Infant, and Neonatal Mortality}
\usage{
dem.chm(
  data,
  age_col,
  deaths_col,
  live_births_col,
  infant_deaths_col,
  neonatal_deaths_col
)
}
\arguments{
\item{data}{A dataframe containing demographic data.}

\item{age_col}{The column name for age groups.}

\item{deaths_col}{The column name for total deaths.}

\item{live_births_col}{The column name for live births.}

\item{infant_deaths_col}{The column name for infant deaths.}

\item{neonatal_deaths_col}{The column name for neonatal deaths.}
}
\value{
A list with Child Mortality, Infant Mortality, and Neonatal Mortality.
}
\description{
This function calculates Child Mortality Rate, Infant Mortality Rate, and Neonatal Mortality Rate.
}
\examples{
demo_data <- data.frame(
  age = c("0-4", "5-9", "10-14"),
  infant_deaths = c(20, 15, 10),
  neonatal_deaths = c(5, 4, 3),
  total_deaths = c(100, 50, 30),
  live_births = c(2000, 2500, 2400)
)
dem.chm(demo_data, age_col = "age", deaths_col = "total_deaths",
                          live_births_col = "live_births",
                          infant_deaths_col = "infant_deaths",
                          neonatal_deaths_col = "neonatal_deaths")
}
