% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetables.R
\name{dem.lifetables}
\alias{dem.lifetables}
\title{Life Table Computations}
\usage{
dem.lifetables(
  data,
  age = "Age",
  nMx = NULL,
  pop = NULL,
  Dx = NULL,
  nqx = NULL
)
}
\arguments{
\item{age}{Numeric vector: age groups to be presented in point format, e.g., 0, 1, 5, 10, ...}

\item{nMx}{Numeric vector (optional): the observed mortality rates in the population. If not provided, \code{pop} and \code{Dx} must be given to compute it.}

\item{pop}{Numeric vector (optional): total population in each age group. Used to compute \code{nMx} if not given.}

\item{Dx}{Numeric vector (optional): deaths occurred in each age group. Used to compute \code{nMx} if not given.}

\item{nqx}{Numeric vector (optional):probability of dying}
}
\value{
A list containing:
\describe{
\item{\code{metrics}}{A list of all calculated life table metrics.}
\item{\code{lifetable}}{A dataframe of the complete life table.}
}
}
\description{
This function computes a complete life table based on given mortality rates or raw data.
It takes a dataframe with the necessary inputs and computes standard life table metrics.
}
\examples{
#Example without moratlity rate
data(gphc2010)
dem.lifetables(data=gphc2010, age = "Age", pop = "Pop", Dx = "Deaths")


}
