% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maternal_Mortality.R
\name{dem.mmr}
\alias{dem.mmr}
\title{Calculate Maternal Mortality Metrics}
\usage{
dem.mmr(data, deaths_col, births_col, pop_women)
}
\arguments{
\item{data}{A dataframe containing demographic data.}

\item{deaths_col}{The column name for maternal deaths.}

\item{births_col}{The column name for live births.}

\item{pop_women}{The column containing the number of women in the reporductive age}
}
\value{
A list with Maternal Mortality Rate (MMR) and Maternal Mortality Ratio (MMR).
}
\description{
This function calculates the Maternal Mortality Rate (MMR) and Maternal Mortality Ratio (MMR).
}
\examples{
demo_data <- data.frame(
  age=c("15-24", "25-34", "35-44"),
  maternal_deaths = c(5, 10, 15),
  live_births = c(5000, 7000, 6000),
  women =c(20000, 22000, 32000)
)
dem.mmr(demo_data, deaths_col = "maternal_deaths", births_col = "live_births", pop_women = "women")
}
