% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemographicModelBase.R
\name{DemographicModelBase}
\alias{DemographicModelBase}
\title{Base Demographic Model Clas}
\value{
A processed data frame with standardized variables for the models to inherit
}
\description{
Base R6 Class that provides common functionality for demographic models
}
\examples{

## ------------------------------------------------
## Method `DemographicModelBase$new`
## ------------------------------------------------

var_mapping <- list(
  region = "region",
  district = "district",
  time = "year",
  age = "age_group",
  sex = "sex",
  population = "population",
  deaths = "deaths",
  in_migration= "in_migration" ,
  out_migration = "out_migration"
)


DemographicModelBase$new(data=testdata, variable_mapping = var_mapping)

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{The model's data}

\item{\code{var_map}}{Variable mapping}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DemographicModelBase-new}{\code{DemographicModelBase$new()}}
\item \href{#method-DemographicModelBase-prepare_model_data}{\code{DemographicModelBase$prepare_model_data()}}
\item \href{#method-DemographicModelBase-get_mapping}{\code{DemographicModelBase$get_mapping()}}
\item \href{#method-DemographicModelBase-clone}{\code{DemographicModelBase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DemographicModelBase-new"></a>}}
\if{latex}{\out{\hypertarget{method-DemographicModelBase-new}{}}}
\subsection{Method \code{new()}}{
Initialize the base model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicModelBase$new(data, variable_mapping)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data frame containing demographic data with variables listed in the variable mapping}

\item{\code{variable_mapping}}{List of variable names mapping to the data frame}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{var_mapping <- list(
  region = "region",
  district = "district",
  time = "year",
  age = "age_group",
  sex = "sex",
  population = "population",
  deaths = "deaths",
  in_migration= "in_migration" ,
  out_migration = "out_migration"
)


DemographicModelBase$new(data=testdata, variable_mapping = var_mapping)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DemographicModelBase-prepare_model_data"></a>}}
\if{latex}{\out{\hypertarget{method-DemographicModelBase-prepare_model_data}{}}}
\subsection{Method \code{prepare_model_data()}}{
Prepare data for modeling
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicModelBase$prepare_model_data(data = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The model's data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DemographicModelBase-get_mapping"></a>}}
\if{latex}{\out{\hypertarget{method-DemographicModelBase-get_mapping}{}}}
\subsection{Method \code{get_mapping()}}{
Get variable mapping
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicModelBase$get_mapping()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DemographicModelBase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DemographicModelBase-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DemographicModelBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
