% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FertilityModel.R
\name{FertilityModel}
\alias{FertilityModel}
\title{Fertility Model}
\description{
R6 Class for modeling fertility rates using Bayesian methods
}
\details{
This class is used to fit a Bayesian model for fertility rates using the brms package.
The model is fitted to a dataset containing births and population data, and can be used to predict future fertility rates.
}
\examples{
# example code
var_mapping <- list(
  region = "region",
  district = "district",
  time = "year",
  age = "age_group",
  sex = "sex",
  population = "population",
  deaths = "deaths",
  in_migration= "in_migration" ,
  out_migration = "out_migration"
)

# Create MCMC settings
mcmc_settings <- McmcSettings$new(
 iterations = 1000,
 warmup = 500,
 chains = 4,
 cores = 4,
 adapt_delta = 0.99,
 max_treedepth = 20,
 seed = 42
)

#initialise the model
fertility_model <- FertilityModel$new(data = testdata, variable_mapping = var_mapping, mcmc_settings = mcmc_settings)

#fit the model using
#fertility_model$fit()

}
\section{Super class}{
\code{\link[DemogAnr:DemographicModelBase]{DemogAnr::DemographicModelBase}} -> \code{FertilityModel}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model}}{Fitted brms model for fertility}

\item{\code{mcmc_settings}}{markov chain monte carlo settings for model fitting and adjusting model performance}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FertilityModel-new}{\code{FertilityModel$new()}}
\item \href{#method-FertilityModel-fit}{\code{FertilityModel$fit()}}
\item \href{#method-FertilityModel-predict_fertility}{\code{FertilityModel$predict_fertility()}}
\item \href{#method-FertilityModel-clone}{\code{FertilityModel$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DemogAnr" data-topic="DemographicModelBase" data-id="get_mapping"><a href='../../DemogAnr/html/DemographicModelBase.html#method-DemographicModelBase-get_mapping'><code>DemogAnr::DemographicModelBase$get_mapping()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DemogAnr" data-topic="DemographicModelBase" data-id="prepare_model_data"><a href='../../DemogAnr/html/DemographicModelBase.html#method-DemographicModelBase-prepare_model_data'><code>DemogAnr::DemographicModelBase$prepare_model_data()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FertilityModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-FertilityModel-new}{}}}
\subsection{Method \code{new()}}{
Initialize the fertility model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FertilityModel$new(data, variable_mapping, mcmc_settings = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data frame containing births and population data}

\item{\code{variable_mapping}}{Variable mapping for model variables}

\item{\code{mcmc_settings}}{MCMC settings for model fitting}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FertilityModel-fit"></a>}}
\if{latex}{\out{\hypertarget{method-FertilityModel-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the fertility model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FertilityModel$fit(formula = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{formula}}{Beysian Formula for the model}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FertilityModel-predict_fertility"></a>}}
\if{latex}{\out{\hypertarget{method-FertilityModel-predict_fertility}{}}}
\subsection{Method \code{predict_fertility()}}{
Predict fertility rates
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FertilityModel$predict_fertility(newdata)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newdata}}{Data frame containing new data for prediction}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FertilityModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FertilityModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FertilityModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
