% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagout.R
\name{flagout}
\alias{flagout}
\title{Title Flag Outliers}
\usage{
flagout(
  data,
  varname,
  item,
  weight = NULL,
  over = NULL,
  z = 3,
  min_n = 20,
  verbose = FALSE
)
}
\arguments{
\item{data}{data to be passed to the function}

\item{varname}{varialble to be analysed}

\item{item}{the first grouping variable, or the item variable if no other grouping is allowed}

\item{weight}{the weight variable, if any}

\item{over}{the other grouping variable, if any}

\item{z}{the z score or the cut of point}

\item{min_n}{the minimum number of observations in a group}

\item{verbose}{if TRUE, print out the summary of flagged outliers and group-level statistics}
}
\value{
a data frame with the original data and additional columns for flagged outliers
}
\description{
This function flags outliers in a dataset based on user defined z-scores or
cut-off points on a normally distributed values.
}
\examples{

# loading required libraries
library(haven)
library(dplyr)

# loading data sets

HH_MOD_I1 <- read_dta("C:/Users/PROXIMITY REPORT/Desktop/Poverty Analysis/Malawi data/IHS5/HH_MOD_I1.dta")

mum<- read_dta("C:/Users/PROXIMITY REPORT/Desktop/Poverty Analysis/Malawi data/IHS5/hh_mod_a_filt.dta")

# preparing the data for analysis
data <- mum |>
 inner_join(HH_MOD_I1, by = "case_id") |>
 mutate(across(where(is.labelled), as_factor),
        logvalue = log(hh_i03/hhsize)) |>
 filter(hh_i01 != "No")

result <- flagout(
data = data,
varname = "logvalue",
item = "hh_i02",
z = 3.5,
verbose = TRUE,
over = c("reside", "region"),
weight = "hh_wgt"
)

}
