% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gphc2010}
\alias{gphc2010}
\title{Example Dataset: gphc2010}
\format{
A data frame with 21 rows and 3 columns:
\describe{
\item{Age}{Age groups expressed as single ages}
\item{Pop}{population from the Ghana 2010 census}
\item{Deaths}{ number of deaths from the Ghana 2010 census}
}
}
\source{
Ghana Statistical Service
}
\usage{
gphc2010
}
\description{
This dataset contains Ghana's 2010 Population and Housing Census (GPHC) data.
}
\keyword{datasets}
