\name{HTRU}
\alias{HTRU}
\docType{data}
\title{
HTRU2
}
\description{
The HTRU2 data
}
\usage{data("HTRU")}
\format{
  A data frame with 17898 observations on the following 9 variables.
  \describe{
    \item{\code{m1}}{a numeric vector}
    \item{\code{m2}}{a numeric vector}
    \item{\code{m3}}{a numeric vector}
    \item{\code{m4}}{a numeric vector}
    \item{\code{m5}}{a numeric vector}
    \item{\code{m6}}{a numeric vector}
    \item{\code{m7}}{a numeric vector}
    \item{\code{m8}}{a numeric vector}
    \item{\code{c}}{a numeric vector}
  }
}
\details{
The HTRU2 data is mainly composed of several pulsar candidate samples, which contains 17898 data points, including the 9 variables. 
}
\source{
The HTRU2 data set is from the UCI database. 
}
\references{
R. J. Lyon, HTRU2, DOI: 10.6084/m9.figshare.3080389.v1.
}
\examples{
data(HTRU)
## maybe str(HTRU) ; plot(HTRU) ...
}
\keyword{datasets}
