% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{last_coef}
\alias{last_coef}
\title{Example Dataset: last_coef}
\format{
A numeric matrix with 5 rows and 3 columns:
\describe{
\item{Column 1}{First set of coefficients.}
\item{Column 2}{Second set of coefficients.}
\item{Column 3}{Third set of coefficients.}
}
}
\source{
Karup-king coefficients for age splitting
}
\usage{
last_coef
}
\description{
These datasets contain coefficients for demographic computations.
}
\keyword{datasets}
