% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MigrationModel.R
\name{MigrationModel}
\alias{MigrationModel}
\title{Migration Model Class}
\description{
R6 Class for modeling both in and out migration using Bayesian methods
}
\section{Super class}{
\code{\link[Dem:DemographicModelBase]{Dem::DemographicModelBase}} -> \code{MigrationModel}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{in_model}}{Fitted brms model for in-migration}

\item{\code{out_model}}{Fitted brms model for out-migration}

\item{\code{mcmc_settings}}{MCMC settings for model fitting}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MigrationModel-new}{\code{MigrationModel$new()}}
\item \href{#method-MigrationModel-fit}{\code{MigrationModel$fit()}}
\item \href{#method-MigrationModel-predict_migration}{\code{MigrationModel$predict_migration()}}
\item \href{#method-MigrationModel-calculate_net_migration}{\code{MigrationModel$calculate_net_migration()}}
\item \href{#method-MigrationModel-print_summaries}{\code{MigrationModel$print_summaries()}}
\item \href{#method-MigrationModel-clone}{\code{MigrationModel$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="Dem" data-topic="DemographicModelBase" data-id="get_mapping"><a href='../../Dem/html/DemographicModelBase.html#method-DemographicModelBase-get_mapping'><code>Dem::DemographicModelBase$get_mapping()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Dem" data-topic="DemographicModelBase" data-id="prepare_model_data"><a href='../../Dem/html/DemographicModelBase.html#method-DemographicModelBase-prepare_model_data'><code>Dem::DemographicModelBase$prepare_model_data()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MigrationModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-MigrationModel-new}{}}}
\subsection{Method \code{new()}}{
Initialize the migration model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MigrationModel$new(data, variable_mapping, mcmc_settings = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MigrationModel-fit"></a>}}
\if{latex}{\out{\hypertarget{method-MigrationModel-fit}{}}}
\subsection{Method \code{fit()}}{
Fit both in and out migration models
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MigrationModel$fit(in_formula = NULL, out_formula = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MigrationModel-predict_migration"></a>}}
\if{latex}{\out{\hypertarget{method-MigrationModel-predict_migration}{}}}
\subsection{Method \code{predict_migration()}}{
Predict migration flows
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MigrationModel$predict_migration(newdata, type = "both")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newdata}}{New data for prediction}

\item{\code{type}}{Type of prediction ("in", "out", or "both")}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MigrationModel-calculate_net_migration"></a>}}
\if{latex}{\out{\hypertarget{method-MigrationModel-calculate_net_migration}{}}}
\subsection{Method \code{calculate_net_migration()}}{
Calculate net migration
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MigrationModel$calculate_net_migration(predictions)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MigrationModel-print_summaries"></a>}}
\if{latex}{\out{\hypertarget{method-MigrationModel-print_summaries}{}}}
\subsection{Method \code{print_summaries()}}{
Print model summaries
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MigrationModel$print_summaries()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MigrationModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MigrationModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MigrationModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
