% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortalityModel.R
\name{MortalityModel}
\alias{MortalityModel}
\title{Mortality Model Class}
\description{
R6 Class for modeling mortality rates using Bayesian methods
}
\examples{
# example code
mcmc_settings <- McmcSettings$new(
iterations = 1000,
warmup = 500,
chains = 4,
cores = 4,
adapt_delta = 0.99,
max_treedepth = 20,
seed = 42
)

# Define variable mapping
var_mapping <- list(
 region = "region",
 district = "district",
 time = "year",
 age = "age_group",
 sex = "sex",
 population = "population",
 deaths = "deaths",
 in_migration= "in_migration" ,
 out_migration = "out_migration"
)
mortality_model <- MortalityModel$new(
data = testdata,
variable_mapping = var_mapping,
mcmc_settings = mcmc_settings
)

#fit the motality model
# mortality_model$fit()

}
\section{Super class}{
\code{\link[Dem:DemographicModelBase]{Dem::DemographicModelBase}} -> \code{MortalityModel}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model}}{Fitted brms model for mortality}

\item{\code{mcmc_settings}}{MCMC settings for model fitting}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MortalityModel-new}{\code{MortalityModel$new()}}
\item \href{#method-MortalityModel-fit}{\code{MortalityModel$fit()}}
\item \href{#method-MortalityModel-predict_mortality}{\code{MortalityModel$predict_mortality()}}
\item \href{#method-MortalityModel-clone}{\code{MortalityModel$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="Dem" data-topic="DemographicModelBase" data-id="get_mapping"><a href='../../Dem/html/DemographicModelBase.html#method-DemographicModelBase-get_mapping'><code>Dem::DemographicModelBase$get_mapping()</code></a></span></li>
<li><span class="pkg-link" data-pkg="Dem" data-topic="DemographicModelBase" data-id="prepare_model_data"><a href='../../Dem/html/DemographicModelBase.html#method-DemographicModelBase-prepare_model_data'><code>Dem::DemographicModelBase$prepare_model_data()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MortalityModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-MortalityModel-new}{}}}
\subsection{Method \code{new()}}{
Initialize the mortality model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MortalityModel$new(data, variable_mapping, mcmc_settings = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data frame with mortality data}

\item{\code{variable_mapping}}{List with variable mappings}

\item{\code{mcmc_settings}}{List with MCMC settings}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MortalityModel-fit"></a>}}
\if{latex}{\out{\hypertarget{method-MortalityModel-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the mortality model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MortalityModel$fit(formula = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{formula}}{Formula for the model}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MortalityModel-predict_mortality"></a>}}
\if{latex}{\out{\hypertarget{method-MortalityModel-predict_mortality}{}}}
\subsection{Method \code{predict_mortality()}}{
Predict mortality rates
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MortalityModel$predict_mortality(newdata)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newdata}}{Data frame with new data}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Predicted mortality rates
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MortalityModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MortalityModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MortalityModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
