% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationProjector.R
\name{PopulationProjector}
\alias{PopulationProjector}
\title{Population Projector Class}
\description{
R6 Class for combining demographic models and projecting population
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fertility_model}}{FertilityModel instance}

\item{\code{mortality_model}}{MortalityModel instance}

\item{\code{migration_model}}{MigrationModel instance}

\item{\code{projections}}{Stored projection results}

\item{\code{projection_variants}}{Different projection scenarios}

\item{\code{simulations}}{Stored simulation results}

\item{\code{credible_intervals}}{Calculated credible intervals}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PopulationProjector-new}{\code{PopulationProjector$new()}}
\item \href{#method-PopulationProjector-project}{\code{PopulationProjector$project()}}
\item \href{#method-PopulationProjector-print_summary}{\code{PopulationProjector$print_summary()}}
\item \href{#method-PopulationProjector-clone}{\code{PopulationProjector$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PopulationProjector-new"></a>}}
\if{latex}{\out{\hypertarget{method-PopulationProjector-new}{}}}
\subsection{Method \code{new()}}{
Initialize the population projector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationProjector$new(
  fertility_model = NULL,
  mortality_model = NULL,
  migration_model = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PopulationProjector-project"></a>}}
\if{latex}{\out{\hypertarget{method-PopulationProjector-project}{}}}
\subsection{Method \code{project()}}{
Project population
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationProjector$project(base_population, horizon_year, n_trajectories = 10)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{base_population}}{Initial population data}

\item{\code{horizon_year}}{Target year for projection}

\item{\code{n_trajectories}}{Number of trajectories to simulate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PopulationProjector-print_summary"></a>}}
\if{latex}{\out{\hypertarget{method-PopulationProjector-print_summary}{}}}
\subsection{Method \code{print_summary()}}{
Print projection summary
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationProjector$print_summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PopulationProjector-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PopulationProjector-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PopulationProjector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
